+++ src/collector/poller/mintscan.rs
@@ -1,13 +1,7 @@
 //! Mintscan poller
 
 use crate::{collector, config, network, prelude::*};
-use datadog;
-use datadog::{send_stream_event, StreamEvent};
-use hostname;
 use mintscan::{Address, Mintscan};
-use std::collections::BTreeMap;
-use std::env;
-use std::time::SystemTime;
 use tendermint::chain;
 use tower::{util::ServiceExt, Service};
 
@@ -63,43 +57,13 @@ impl Poller {
         };
 
         let mut last_signed_height = None;
-
         if let Some(addr) = &self.validator_addr {
             match self.client.validator_uptime(addr).await {
                 Ok(uptime) => {
                     dbg!(&uptime);
                     last_signed_height = Some(uptime.latest_height.into());
-                    dbg!(uptime.uptime.len());
-                    if uptime.uptime.len() >= 2 {
-                        let dd_api_key = env::var("DD_API_KEY").unwrap();
-                        let hostname = hostname::get().unwrap();
-                        let mut ddtags = BTreeMap::new();
-                        ddtags.insert("env".to_owned(), "staging".to_owned());
-                        let stream_event = StreamEvent {
-                            aggregation_key: None,
-                            alert_type: Some(datadog::AlertType::Error),
-                            date_happened: Some(SystemTime::now()),
-                            device_name: None,
-                            hostname: Some(hostname.to_string_lossy().to_string()),
-                            priority: Some(datadog::Priority::Normal),
-                            related_event_id: None,
-                            tags: Some(ddtags),
-                            // Text field must contain @pagerduty to trigger alert
-                            text: format!("@pagerduty missed blocks alert: {:?}", &uptime),
-                            title: "missed blocks alert test".to_owned(),
-                        };
 
-                        // send stream event to datadog which forwards to pagerduty
-                        let stream_event = send_stream_event(&stream_event, dd_api_key).await;
-                        match stream_event {
-                            Ok(()) => {
-                                dbg!("event sent to datadog");
-                            }
-                            Err(_err) => {
-                                warn!("unable to sent event to datadog");
-                            }
-                        }
-                    }
+                    dbg!(uptime.uptime.len());
                 }
                 Err(err) => {
                     warn!(
