+++ src/network/tendermint.rs
@@ -1,12 +1,20 @@
 //! Tendermint network types
 
+
+use datadog;
+use datadog::{send_stream_event, StreamEvent};
+use hostname;
 use super::Id;
 use crate::{
     message::{Envelope, Message},
     monitor::{net_info::Peer, status::ChainStatus},
     prelude::*,
+    collector::PollEvent
 };
 use serde::Serialize;
+use std::collections::BTreeMap;
+use std::env;
+use std::time::SystemTime;
 
 /// Tendermint network
 #[derive(Debug, Clone)]
@@ -64,6 +72,45 @@ impl Network {
                 Message::Validator(ref validator_info) => self.update_validator(validator_info),
             }
         }
+    }
+
+    /// Handle incoming poll event
+    pub async fn handle_poll_event(&mut self, poll_event: PollEvent) {
+        dbg!(&poll_event);
+        let last_signed_height = poll_event.last_signed_height.unwrap();
+        let page_threshold = last_signed_height + 20;
+        let current_height = poll_event.current_height;
+        if current_height > page_threshold {
+            let dd_api_key = env::var("DD_API_KEY").unwrap();
+            let hostname = hostname::get().unwrap();
+            let mut ddtags = BTreeMap::new();
+            ddtags.insert("env".to_owned(), "staging".to_owned());
+            let stream_event = StreamEvent {
+                aggregation_key: None,
+                alert_type: Some(datadog::AlertType::Error),
+                date_happened: Some(SystemTime::now()),
+                device_name: None,
+                hostname: Some(hostname.to_string_lossy().to_string()),
+                priority: Some(datadog::Priority::Normal),
+                related_event_id: None,
+                tags: Some(ddtags),
+                // Text field must contain @pagerduty to trigger alert
+                text: format!("@pagerduty cosmon poll event: {:?}", &poll_event),
+                title: "cosmon poll event".to_owned(),
+            };
+
+            // send stream event to datadog which forwards to pagerduty
+            let stream_event = send_stream_event(&stream_event, dd_api_key).await;
+            match stream_event {
+                Ok(()) => {
+                    dbg!("event sent to datadog");
+                }
+                Err(_err) => {
+                    warn!("unable to sent event to datadog");
+                }
+            }
+        }
+
     }
 
     /// Update information about a particular node
